Separate out hash_test.cpp source for each hash API
